CC/requirements.txt
absl-py==0.12.0
aniso8601==8.0.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dbus-python==1.2.16
distlib==0.3.4
distro-info==1.0
filelock==3.7.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==2.10
importlib-metadata==1.6.0
ims==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
jsonfy==0.4
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
more-itertools==4.2.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
Pillow==8.2.0
platformdirs==2.5.2
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-apt==2.2.1
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.18
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.14.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.0.0